moved useTooltip to be along side Tooltip
