+++ src/containers/Ledgers/index.tsx
@@ -12,7 +12,7 @@ import { Ledger, ValidatorResponse } from './types'
 import { useAnalytics } from '../shared/analytics'
 import NetworkContext from '../shared/NetworkContext'
 import { useIsOnline } from '../shared/SocketContext'
-import { TooltipProvider } from './useTooltip'
+import { TooltipProvider } from '../shared/components/Tooltip'
 import { SelectedValidatorProvider } from './useSelectedValidator'
 
 const FETCH_INTERVAL_MILLIS = 5 * 60 * 1000
