+++ src/containers/shared/components/Tooltip/index.ts
@@ -1,0 +1,2 @@
+export * from './Tooltip'
+export * from './useTooltip'
